daemon/server/httputils: move WriteLogStream to its own package
